flask/01_celery_flask/requirements.txt
flask==3.0.1
python-dotenv==1.0.1
celery==5.3.6
mysql-connector-python==8.3.0
redis==5.0.1
gunicorn==21.2.0
marshmallow==3.20.2